Update PopoverItem,  Add PopoverDirection to exports
