+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [0.0.2] - 18.01.2021.
+
+* Update `PopoverItem`.
+* Add `PopoverDirection` to exports.
+* Add documentation.
+
 ## [0.0.1] - 10.01.2021.
 
 * Initial release.
