+++ lib/popover.dart
@@ -1,1 +1,2 @@
 export 'src/popover.dart';
+export 'src/popover_direction.dart';
