+++ lib/src/popover_direction.dart
@@ -1,3 +1,4 @@
+/// Popover direction
 enum PopoverDirection {
   top,
   bottom,
