+++ pubspec.yaml
@@ -1,6 +1,6 @@
 name: popover
-description: Popover for Flutter.
-version: 0.0.1
+description: A popover is a transient view that appears above other content onscreen when you tap a control or in an area.
+version: 0.0.2
 homepage: https://github.com/minikin/popover
 
 environment:
